+++ install/migrations/0047_add_thumbnail_media_id_to_downloads_table.sql
@@ -1,0 +1,1 @@
+ALTER TABLE {prefix}downloads ADD COLUMN thumbnail_media_id INT UNSIGNED NULL AFTER media_id;